requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
ansible==6.7.0
ansible-core==2.13.13
ansible-vault==2.1.0
anyio==4.4.0
astunparse==1.6.3
attrs==23.2.0
backoff==2.2.1
base58==2.1.1
bittensor @ file:///D:/04_Project/10_Bittensor/02_src/bittensor
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
configparser==7.0.0
construct==2.10.68
construct-typing==0.5.6
cryptography==42.0.8
Cython==3.0.10
cytoolz==0.12.3
ddt==1.6.0
decorator==5.1.1
dm-tree==0.1.8
docstring_parser==0.16
ecdsa==0.19.0
eth-hash==0.7.0
eth-keys==0.5.1
eth-typing==4.4.0
eth-utils==2.2.2
etils==1.9.1
fastapi==0.110.3
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.0
fuzzywuzzy==0.18.0
gast==0.5.4
google-pasta==0.2.0
googleapis-common-protos==1.63.1
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
immutabledict==4.2.0
importlib_resources==6.4.0
jax==0.4.29
jaxlib==0.4.29
Jinja2==3.1.4
jsonalias==0.1.1
kagglehub==0.2.6
keras==3.3.3
keras-core==0.1.7
keras-cv==0.9.0
Levenshtein==0.25.1
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
more-itertools==10.3.0
msgpack==1.0.8
msgpack-numpy-opentensor==0.5.0
multidict==6.0.5
munch==2.5.0
namex==0.0.8
nest-asyncio==1.6.0
netaddr==1.3.0
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.1
pandas==2.2.2
password-strength==0.0.3.post2
promise==2.3
protobuf==4.25.3
psutil==5.9.8
py==1.11.0
py-bip39-bindings==0.1.11
py-ed25519-zebra-bindings==1.0.1
py-sr25519-bindings==0.2.0
pyaes==1.6.1
pyarrow==16.1.0
pyasn1==0.6.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.25.1
python-statemachine==2.1.2
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.4
regex==2024.5.15
requests==2.32.3
resolvelib==0.8.1
retry==0.9.2
rich==13.7.1
rsa==4.9
scalecodec==1.2.7
scipy==1.13.1
setuptools==70.0.0
shtab==1.6.5
simple_parsing==0.1.5
six==1.16.0
sniffio==1.3.1
solana==0.34.3
solders==0.21.0
starlette==0.37.2
substrate-interface==1.7.5
Telethon==1.36.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-datasets==4.9.6
tensorflow-intel==2.16.1
tensorflow-metadata==1.15.0
termcolor==2.4.0
toml==0.10.2
toolz==0.12.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.30.0
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
